Moving parsing over to YAML based files and added rst generation.
