+++ .gitignore
@@ -8,6 +8,7 @@ build
 \#*\#
 *.swp
 ipfabric/
+docs/source/kb.rst
 docs/source/solutions/ipfabric/
 docs/source/_static/images/3_clos_topology.jpg
 docs/source/_static/images/bwc_components.jpg
@@ -15,3 +16,5 @@ docs/source/_static/images/ipfabric_topology.jpg
 docs/source/_static/images/ztp_topology.jpg
 docs/source/_static/images/ztp_usecase_1.jpg
 docs/source/_static/images/ztp_usecase_2.jpg
+
+pandoc*
